==> prepare/02A-id_to_publication_year-PMC.ipynb <==
# PMID to Year Map -- Pubmed Central

import pandas as pd
from tqdm import tqdm

!wget -O ../data/PMC-ids-csv.gz ftp://ftp.ncbi.nlm.nih.gov/pub/pmc/PMC-ids.csv.gz

pmids = []
with open('../data/pmid_list.txt', 'r') as fin:
    for line in fin.readlines():
        pmids.append(line.strip())
print('{:,}'.format(len(pmids)))

df = pd.read_csv('../data/PMC-ids.csv.gz')

df.head()

df.dtypes

print('{:,}'.format(len(df)))
print('{:,}'.format(df['PMID'].count()))

df = df.dropna(subset=['PMID'])
df['PMID'] = df['PMID'].astype(int).astype(str)

df.head(2)

pmid_to_year = df.set_index('PMID')['Year'].to_dict()

mapped = set(pmid_to_year.keys())
no_map = set(pmids) - mapped
print('{:,}'.format(len(no_map)))

import pickle
pickle.dump(pmid_to_year, open('../data/pmid_to_year_PMC.pkl', 'wb'))
