==> prepare/02B-id_to_publication_year-Europmc.ipynb <==
# PMID to Year Map -- Europepmc

https://europepmc.org/downloads

import os
import pandas as pd
from tqdm import tqdm
import gzip
import xml.etree.ElementTree as ET

!wget -O ../data/PMCLiteMetadata.tgz http://europepmc.org/ftp/pmclitemetadata/PMCLiteMetadata.tgz

!tar -xvzf ../data/PMCLiteMetadata.tgz -C ../data

import xml.etree.ElementTree as ET

class XML2DataFrame:

    def __init__(self, xml_file):
        self.root = ET.parse(xml_file).getroot()

    def parse_root(self, root):
        """Return a list of dictionaries from the text
         and attributes of the children under this XML root."""
        return [self.parse_element(child) for child in root.getchildren()]

    def parse_element(self, element):
        """ Collect {key:attribute} and {tag:text} from thie XML
         element and all its children into a single dictionary of strings."""
        parsed = {c.tag: c.text for c in element.getchildren()}
        return parsed
 
    def process_data(self):
        """ Initiate the root XML, parse it, and return a dataframe"""
        structure_data = self.parse_root(self.root)
        return pd.DataFrame(structure_data)


frames = []
base = '../data/out'
files = sorted([f for f in os.listdir(base) if f.endswith('.xml')], key=lambda x: int(x.split('.')[1]))

for file in tqdm(files):
    xml2df = XML2DataFrame(os.path.join(base, file))
    xml_dataframe = xml2df.process_data()
    frames.append(xml_dataframe)

result = pd.concat(frames)

result.head(2)

print('{:,}'.format(len(result)))

pmid_mapper = result.set_index('pmid')['pubYear'].to_dict()

import pickle
prev_no_map = pickle.load(open('../data/no_map_PMC.pkl', 'rb'))
print('{:,}'.format(len(prev_no_map)))

mapped = set(pmid_mapper.keys())
new_no_map = prev_no_map - mapped
print('{:,}'.format(len(new_no_map)))

pickle.dump(pmid_mapper, open('../data/pmid_to_year_Eur.pkl', 'wb'))
pickle.dump(new_no_map, open('../data/no_map_Eur.pkl', 'wb'))
