==> prepare/02C-id_to_publication_year-NLM_Baseline.ipynb <==
import os
import pandas as pd
from tqdm import tqdm
import gzip
import xml.etree.ElementTree as ET

Data downloaded from:  
ftp://ftp.ncbi.nlm.nih.gov/pubmed/baseline/

Run `./download_baseline.sh` script first before running this noteboook

def get_child_tag(child, tag):
    for c in child.getchildren():
        if c.tag == tag:
            return c

def get_year_from_article(article):
    journal = get_child_tag(article, 'Journal')
    issue = get_child_tag(journal, 'JournalIssue')
    pub_date = get_child_tag(issue, 'PubDate')
    year = get_child_tag(pub_date, 'Year')
    
    if year is not None:
        return year.text

def get_year_from_pubmed(pubmed_data):
    history = get_child_tag(pubmed_data, 'History')   
    for child in history.getchildren():
        if child.tag == 'PubMedPubDate' and child.items()[0][1] == 'pubmed':
            year = get_child_tag(child, 'Year')
    if year is not None:
        return year.text

def get_pmid_year(pubmed_article):
    medline_cit = get_child_tag(pubmed_article, 'MedlineCitation')
    pubmed_data = get_child_tag(pubmed_article, 'PubmedData')
    
    pmid = get_child_tag(medline_cit, 'PMID')
    try:
        year = get_year_from_pubmed(pubmed_data)
    except:
        article = get_child_tag(medline_cit, 'Article')
        year = get_year_from_article(article)
        
    if pmid is not None:
        pmid = pmid.text
    
    return pmid, year
    

base = '../data/baseline/'
files = [f for f in os.listdir(base) if f.endswith('.xml.gz')]

print(files[0])

# Last 4 characters before .xml indicate file's order
files = sorted(files, key = lambda f: int(f.split('.')[0][-4:]))

len(files)

# This is a stupidly-slow problem of opening g-zipped files... 
# Could be sped up massively via parallel processing

import sys
sys.path.append('../../hetnet-ml/src')
from parallel import parallel_process

def get_id_to_year_map(file):
    id_to_year = {}
    tree = ET.parse(gzip.open(os.path.join(base, file)))
    root = tree.getroot()

    for cit in root.getchildren():
        pmid, year = get_pmid_year(cit)
        id_to_year[pmid] = year
    return id_to_year

id_to_year = {}
results = parallel_process(files, get_id_to_year_map, n_jobs=32, front_num=0)

for r in results:
    id_to_year.update(r)
    
    
print('{:,}'.format(len(id_to_year)))
id_to_year_filt = {k:v for k, v, in id_to_year.items() if v is not None}
print('{:,}'.format(len(id_to_year_filt)))

import pickle
prev_no_map = pickle.load(open('../data/no_map_Eur.pkl', 'rb'))

still_no_map = set(prev_no_map) - set(id_to_year.keys())
print('{:,}'.format(len(still_no_map)))

pickle.dump(id_to_year, open('../data/pmid_to_year_NLM.pkl', 'wb'))
pickle.dump(still_no_map, open('../data/no_map_NLM.pkl', 'wb'))
