==> prepare/03-umls_cui_to_mesh_descriptorID.ipynb <==
import pickle
import pandas as pd

col_names = ['CUI', 'LAT', 'TS', 'LUI', 'STT', 'SUI', 'ISPREF', 'AUI', 'SAUI',
             'SCUI', 'SDUI', 'SAB', 'TTY', 'CODE', 'STR', 'SRL', 'SUPPRESS', 'CVF']

df = pd.read_csv('../2017AA-full/2017AA/META/MRCONSO.RRF', header=None, sep='|')

print(len(df))
print(df[18].count())

# Fix the fact that rows end with a pipe, producing an extra entirely empty column
df = df.iloc[:, :-1]

df.columns = col_names

df.head(5)

msh_rows = df.query('SAB == "MSH"')

grpd = msh_rows.groupby('CUI')
cui_to_msh = dict()
for label, df in grpd:
    cui_to_msh[label] = list(set(df['SDUI']))
    

print(len(cui_to_msh))

# There are a few with multiple mappings...
for v in cui_to_msh.values():
    if len(v) > 1:
        print(v)

pickle.dump(cui_to_msh, open('../data/UMLS-CUI_to_MeSH-Descripctor.pkl', 'wb'))

## Get a quick and dirty name map for the 1to1 id to values

cui_to_msh_1t1 = {k: v[0] for k, v in cui_to_msh.items() if len(v) == 1}

Just take any mapping... we'll get more "Correct" names for as many as we can later

cuis = list(cui_to_msh_1t1.keys())
to_name = msh_rows.query('CUI in @cuis')
msh_to_name = to_name.drop_duplicates(subset=["SDUI"]).set_index('SDUI')['STR'].to_dict()

pickle.dump(msh_to_name, open('../data/MeSH_to_name_quick_n_dirty.pkl', 'wb'))
