==> prepare/04-Parse_MeSH_data.ipynb <==
!wget -O ../data/supp2018.xml ftp://nlmpubs.nlm.nih.gov/online/mesh/MESH_FILES/xmlmesh/supp2018.xml
!wget -O ../data/desc2018.xml ftp://nlmpubs.nlm.nih.gov/online/mesh/MESH_FILES/xmlmesh/desc2018.xml

import pickle
import xml.etree.ElementTree as ET
from collections import defaultdict

## Go from MeSH Descriptor IDs or MeSH Supplemental IDs to name

root = ET.parse('../data/desc2018.xml').getroot()
root1 = ET.parse('../data/supp2018.xml').getroot()

id_to_name = dict()
for descriptor in root.getchildren():
    for child in descriptor.getchildren():
        if child.tag == 'DescriptorUI':
            uid = child.text
        if child.tag == 'DescriptorName':
            name = child.getchildren()[0].text
    id_to_name[uid] = name

for supplement in root1.getchildren():
    for child in supplement.getchildren():
        if child.tag == 'SupplementalRecordUI':
            uid = child.text
        if child.tag == 'SupplementalRecordName':
            name = child.getchildren()[0].text
    id_to_name[uid] = name

len(id_to_name)

pickle.dump(id_to_name, open('../data/MeSH_DescUID_to_Name.pkl', 'wb'))

## Also get tree-values for childern

id_to_treenumbs = defaultdict(list)
for descriptor in root.getchildren():
    for child in descriptor.getchildren():
        if child.tag == 'DescriptorUI':
            uid = child.text
        if child.tag == 'TreeNumberList':
            for tn in child.getchildren():
                id_to_treenumbs[uid].append(tn.text)

pickle.dump(id_to_treenumbs, open('../data/MeSH_DescUID_to_TreeNumbs.pkl', 'wb'))
