==> prepare/05-MeSH_ID_to_name_via_UMLS.ipynb <==
## Go from MeSH IDs to Names via UMLS

In the notebook `03-umls_cui_to_mesh_descriptorID` we got a map from UMLS CUI to MeSH IDs.  We also grabbed some concept names in case we'd need them later.  However, there's no guarnetee that these are the 'preferred' names, so here we'll go a bit deeper and try to get the preferred ones first

import pickle
import pandas as pd

import sys
sys.path.append('../tools/')
import load_umls

conso = load_umls.open_mrconso()

conso.head(2)

msh_rows = conso.query('SAB == "MSH"')

TTY values give hints to what are the preferred naames...  
Primarily, MH and NM are very close to a 1 to 1 relationship of id to value

See [abbreviations here](https://www.nlm.nih.gov/research/umls/knowledge_sources/metathesaurus/release/abbreviations.html) for meanings of all the abbreviations

for tty in ['PXQ', 'PEP', 'PCE', 'HT', 'QEV',  'NM', 'MH']:

    print('Ratio of names to unique IDs for TYY == "{}": '.format(tty), end='')
    q_res = msh_rows.query('LAT == "ENG" and TTY == @tty')
    print('{}'.format(q_res.shape[0] / q_res['SDUI'].nunique()))

msh_to_name = {}

# Least to most important / redundant
for tty in ['PXQ', 'PEP', 'PCE', 'HT', 'QEV',  'NM', 'MH']:
    q_res = msh_rows.query('LAT == "ENG" and TTY == @tty')
    msh_to_name.update(q_res.set_index('SDUI')['STR'].to_dict())
len(msh_to_name) == msh_rows['SDUI'].nunique()

Add in the old mappings, (overwritten by the new ones), in case any happened to be missed

msh_to_name_old = pickle.load(open('../data/MeSH_to_name_quick_n_dirty.pkl', 'rb'))

msh_to_name_final = {**msh_to_name_old, **msh_to_name}

print('Quick and Dirty MeSH concepts mapped to names: {:,}'.format(len(msh_to_name_old)))
print('MeSH concepts mapped to names now: {:,}'.format(len(msh_to_name)))
print('Total MeSH concepts with mapped names: {:,}'.format(len(msh_to_name_final)))

pickle.dump(msh_to_name_final, open('../data/MeSH_id_to_name_via_UMLS.pkl', 'wb'))
